Split more functionality out of loadPyodide into other files
